users/views.py
# users/views.py - Profile Sync API

from rest_framework import viewsets, status
from rest_framework.decorators import action
from rest_framework.response import Response
from rest_framework.permissions import IsAuthenticated
from django.contrib.auth import get_user_model

from events.team_serializers import ProfileSyncSerializer
from core.models import DomainActivity
from core.serializers import DomainActivitySerializer

User = get_user_model()

from google.auth.transport import requests as google_requests
from google.oauth2 import id_token
from rest_framework_simplejwt.tokens import RefreshToken
from rest_framework.views import APIView
from rest_framework.permissions import AllowAny



from events.team_serializers import ProfileSyncSerializer
from .serializers import UserSerializer


class UserViewSet(viewsets.ReadOnlyModelViewSet):
    """
    Standard User API
    """
    queryset = User.objects.all()
    serializer_class = UserSerializer
    permission_classes = [IsAuthenticated]

    def get_queryset(self):
        # Users can only see themselves by default or public profiles?
        # For now, let's restrict list to empty or handle searching later.
        if self.action == 'list':
            return User.objects.none()
        return super().get_queryset()

    @action(detail=False, methods=['get'])
    def me(self, request):
        """
        GET /api/users/me/
        Return current user info
        """
        serializer = self.get_serializer(request.user)
        return Response(serializer.data)

    @action(detail=True, methods=['get'])
    def accomplishments(self, request, pk=None):
        """
        GET /api/users/{pk}/accomplishments/
        Returns public domain activities (Volunteering, Attendance, etc.)
        """
        user = self.get_object()
        activities = DomainActivity.objects.filter(
            actor=user,
            visibility='public',
            status='active'
        ).order_by('-timestamp')

        # We need a serializer for DomainActivity.
        # Assuming one exists or I'll create a simple inline one if needed?
        # Let's check core/serializers.py first. Assuming it exists.
        serializer = DomainActivitySerializer(activities, many=True)
        return Response(serializer.data)


class ProfileViewSet(viewsets.GenericViewSet):
    """
    API for managing user profile data

    Competitive Feature: Profile auto-fill for event registration (parity with devnovate)
    Trust Advantage: Privacy controls + audit trail via snapshots
    """
    permission_classes = [IsAuthenticated]
    serializer_class = ProfileSyncSerializer

    def get_object(self):
        return self.request.user

    @action(detail=False, methods=['get'], url_path='sync-data')
    def sync_data(self, request):
        """
        Get user's profile data for event registration auto-fill

        GET /api/users/profile/sync-data/

        Returns profile fields if allow_profile_autofill=True
        """
        user = request.user

        if not user.allow_profile_autofill:
            return Response({
                'autofill_enabled': False,
                'message': 'Profile auto-fill is disabled. Enable it in settings to use this feature.'
            })

        serializer = ProfileSyncSerializer(user)
        return Response({
            'autofill_enabled': True,
            'profile_data': serializer.data
        })

    @action(detail=False, methods=['patch'], url_path='update-sync')
    def update_sync(self, request):
        """
        Update profile sync data

        PATCH /api/users/profile/update-sync/
        Body: {field_name: value, ...}
        """
        user = request.user
        serializer = ProfileSyncSerializer(user, data=request.data, partial=True)
        serializer.is_valid(raise_exception=True)
        serializer.save()

        return Response({
            'message': 'Profile updated successfully',
            'profile_data': serializer.data
        })

    @action(detail=False, methods=['post'], url_path='toggle-autofill')
    def toggle_autofill(self, request):
        """
        Enable/disable profile auto-fill

        POST /api/users/profile/toggle-autofill/
        Body: {"enabled": true/false}
        """
        user = request.user
        enabled = request.data.get('enabled', True)

        user.allow_profile_autofill = enabled
        user.save()

        return Response({
            'autofill_enabled': enabled,
            'message': f"Profile auto-fill {'enabled' if enabled else 'disabled'}"
        })


class GoogleLoginView(APIView):
    permission_classes = [AllowAny]

    def post(self, request):
        token = request.data.get('id_token')
        if not token:
            return Response({'error': 'ID Token is required'}, status=status.HTTP_400_BAD_REQUEST)

        try:
            # Verify the token
            # CLIENT_ID is optional here if we trust the issuer, but better to enforce it if known.
            # For now, we'll accept any valid Google token as requested in the plan (using placeholder if needed, but the lib handles validation).
            # Specify the CLIENT_ID of the app that accesses the backend:
            # id_info = id_token.verify_oauth2_token(token, google_requests.Request(), "YOUR_GOOGLE_CLIENT_ID")

            # Since we don't have the client ID yet, we can verify without audience check or just verify signature.
            # But specific audience check is recommended.
            # We will use strict verification but allow any audience for now or expect it in env.
            id_info = id_token.verify_oauth2_token(token, google_requests.Request())

            # if 'aud' not in id_info or id_info['aud'] != CLIENT_ID:
            #     raise ValueError('Could not verify audience.')

            email = id_info.get('email')
            if not email:
                return Response({'error': 'Email not found in token'}, status=status.HTTP_400_BAD_REQUEST)

            # Check if user exists
            try:
                user = User.objects.get(email=email)
            except User.DoesNotExist:
                # Create user
                username = email.split('@')[0]
                # Ensure unique username
                if User.objects.filter(username=username).exists():
                    username = f"{username}_{User.objects.count()}"

                user = User.objects.create_user(
                    username=username,
                    email=email,
                    password=None, # Unusable password
                )
                user.is_active = True
                user.save()

            # Generate tokens
            refresh = RefreshToken.for_user(user)

            return Response({
                'refresh': str(refresh),
                'access': str(refresh.access_token),
                'user': {
                    'email': user.email,
                    'username': user.username,
                    'id': user.id
                }
            })

        except ValueError as e:
            return Response({'error': f'Invalid token: {str(e)}'}, status=status.HTTP_400_BAD_REQUEST)

