requirements.txt
Django==4.2.8
djangorestframework==3.14.0
gunicorn==21.2.0
psycopg2-binary==2.9.9
reportlab==4.0.8
qrcode[pil]==7.4.2
Pillow==10.1.0
celery==5.3.6
redis==5.0.1
python-dotenv==1.0.0
cryptography==41.0.7
pymysql==1.1.0
django-extensions==3.2.3
django-cors-headers==4.3.1
djangorestframework-simplejwt==5.3.1
supabase==2.3.0
PyJWT==2.8.0
